+++ fs24bot3/Models/SQL.cs
@@ -65,7 +65,7 @@ public class SQL
         [PrimaryKey]
         [AutoIncrement]
         public int Id { get; set; }
-        public uint RemindDate { get; set; }
+        public long RemindDate { get; set; }
         public string Nick { get; set; }
         public string Channel { get; set; }
         public string Message { get; set; }
